feat(ci view): add `Ctrl+A` shortcut to run all manual jobs in current job's stage and "later"
Some pipelines needs all jobs passed to merge an MR. And they might have
multiple manual jobs. So, to run all the jobs a user must perform
routine actions to trigger every job. That is not user friendly.

Introduce the `Ctrl+A` shortcut to trigger all manual jobs in the
"current" focused stage and all stages after the current.