fix(ci view): job is still cancelled even when ESC pressed on confirmation
When a user tries to cancel a job a confirmation dialog is shown. If the
user presses ESC to interrupt a further process the dialog is closed but
the job is still canceled. That is not an expected behaviour. Usually,
canceling a dialog is equivalent to "No" answer.

Fix that.