fix(ci trace): job log/trace is cut
When user explores jobs' traces of a single pipeline, one job might have
long trace, another is not. It to open a long trace first then a short
one then the short one is not shown.

In the code, it appers, the view trace `runTrace` function is intended
to be called once in the glab's lifecycle but it is not always true.

Fix that.