fix(ci view): last pipeline is shown for a commit even a tag is specified explicitly
In some cases pipeline for a commit can be run multiple time even after
tag on that commit is set. So, a user might want to specify a tag
explicitly (`glab ci view -b TAG`) to view a tag's pipeline. But the
LAST pipeline for the commit is shown but not the specified TAG's.

Fix that. Use a pipeline for the specified tag.